adds Games route with invitation, sonner for notification and socket listener on root
