+++ frontend/package.json
@@ -18,7 +18,8 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.14.2",
     "roughjs": "^4.6.5",
-    "socket.io-client": "^4.7.2"
+    "socket.io-client": "^4.7.2",
+    "sonner": "^1.2.2"
   },
   "devDependencies": {
     "@fontsource/montserrat": "^5.0.8",
