+++ frontend/src/routes/index.ts
@@ -7,6 +7,7 @@ export { default as Home } from "./Home";
 export { default as ChatNew } from "./ChatNew";
 export { default as Leaderboard } from "./Leaderboard";
 export { default as Games } from "./Games";
+export { default as Game } from "./Game";
 export { default as Settings } from "./Settings";
 export { default as ErrorBoundary } from "./ErrorBoundary";
 export { default as Channels } from "./Channels";
